Update zrml-asset-router (#1321)
* Update zrml-primitives

* Partially update asset-router

* Finalize logic adjustments in asset-router

* Make asset-router tests compilable

* Correct Inspect routing for market assets in Currencies

* Directly invoke Inspect API for Currencies

* Add tests for remaining Unbalances functions

+++ primitives/src/constants/mock.rs
@@ -191,7 +191,7 @@ parameter_type_with_key! {
 
 // System
 parameter_types! {
-    pub const BlockHashCount: u64 = 250;
+    pub const BlockHashCount: u32 = 250;
 }
 
 // Time
